machine-learning/deployed/requirements.txt
fastapi
uvicorn
pillow
torch==1.13.1
numpy
pandas
